:recycle: Move `findhooks` tests to `tests.unit.templates`
